API

Types

GeometricAlgebra.Blade — Type

struct Blade{T<:AbstractFloat} <: AbstractBlade{T}

Blade represented with the floating-point precision of type T. The norm and orientation of a Blade are encoded by its volume. The norm of a Blade is equal to abs(volume) and the orientation of a Blade relative to its basis is equal to sign(volume).

Notes

The grade of a Blade type is greater than 0 and less than the dimension of the space that the blade is embedded in

GeometricAlgebra.Pseudoscalar — Type

struct Pseudoscalar{T<:AbstractFloat} <: AbstractBlade

Pseudoscalar (an $(n-1)$-blade) represented with the floating-point precision of type T. The basis for a Pseudoscalar is the standard basis for $R^n$. The norm and orientation of a Pseudoscalar are encoded in its value. The norm of a Pseudoscalar is equal to abs(value) and the orientation of a Pseudoscalar relative to the standard basis for $R^n$ is equal to sign(value).

GeometricAlgebra.Scalar — Type

struct Scalar{T<:AbstractFloat} <: AbstractScalar{T}

Scalar (0-blade) represented with the floating-point precision of type T. The basis and volume of a Scalar are 1 and the value of the Scalar, respectively.

GeometricAlgebra.One — Type

struct One{T<:AbstractFloat} <: AbstractScalar{T}

Multiplicative identity for a geometric algebra (extended from a real vector space of arbitrary dimension).

GeometricAlgebra.Zero — Type

struct Zero{T<:AbstractFloat} <: AbstractScalar{T}

Additive identity for a geometric algebra (extended from a real vector space of arbitrary dimension).

Basic Operations

TODO

Missing docstring.

Missing docstring for GeometricAlgebra.inverse. Check Documenter's build log for details.

Base.reverse — Function

reverse(M::AbstractMultivector)::AbstractMultivector

Compute the reverse of M.

GeometricAlgebra.dual — Function

dual(M::AbstractMultivector)::AbstractMultivector

Compute the dual of M (relative to the pseudoscalar of the geometric algebra that M is an element of).

Unary Operations

TODO

Type Hierarchy

Abstract Types

GeometricAlgebra.AbstractMultivector — Type

AbstractMultivector{<:AbstractFloat}

Supertype for all multivector types.

Interface

Properties

dim(M::AbstractMultivector)::Int

grades(M::AbstractMultivector)::Vector{Int}

blades(M::AbstractMultivector)::Vector{<:AbstractBlade}

norm(M::AbstractMultivector)::AbstractFloat

getindex(M::AbstractMultivector, k::Int)::Vector{<:AbstractBlade}

Unary Operations

-(M::AbstractMultivector)::AbstractMultivector

reverse(M::AbstractMultivector)::AbstractMultivector

dual(M::AbstractMultivector)::AbstractMultivector

Binary Operations

+(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector

-(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector

*(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector

/(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector

wedge(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector
∧(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector

contract_left(M::AbstractMultivector,
              N::AbstractMultivector)::AbstractMultivector
<(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector

contract_right(M::AbstractMultivector,
               N::AbstractMultivector)::AbstractMultivector
>(M::AbstractMultivector, N::AbstractMultivector)::AbstractMultivector

dot(M::AbstractMultivector, N::AbstractMultivector;
    left=true)::AbstractMultivector
⋅(M::AbstractMultivector, N::AbstractMultivector;
  left=true)::AbstractMultivector

project(M::AbstractMultivector, B::AbstractBlade)::AbstractMultivector

Comparison Functions

==(M::AbstractMultivector, N::AbstractMultivector)::Bool
isequal(M::AbstractMultivector, N::AbstractMultivector)::Bool

≈(M::AbstractMultivector, N::AbstractMultivector)::Bool
isapprox(M::AbstractMultivector, N::AbstractMultivector)::Bool

Utility Functions

convert(::Type{T}, M::AbstractMultivector) where {T<:AbstractFloat}

Implementation

The return value of all methods should preserve the precision of its AbstractMultivector arguments (when possible).

GeometricAlgebra.AbstractBlade — Type

AbstractBlade{<:AbstractFloat}

Supertype for all blade types.

Interface

Properties

grade(B::AbstractBlade)::Int

basis(B::AbstractBlade)::Matrix

volume(B::AbstractBlade)::AbstractFloat

sign(B::AbstractBlade)::Int8

Operations

dual(B::AbstractBlade, C::AbstractBlade)::AbstractBlade

inv(B::AbstractBlade)::AbstractBlade

Implementation

For the AbstractBlade type, the volume of a blade should encode both norm and orientation information.

The return value of all methods should preserve the precision of its AbstractBlade arguments (when possible).

GeometricAlgebra.AbstractScalar — Type

AbstractScalar{<:AbstractFloat}

Supertype for all scalar types.

Interface

Note: the return value of all methods should preserve the precision of its AbstractScalar arguments (when possible).

Properties

value(B::AbstractScalar{T})::T where {T<:AbstractFloat}