GeometricAlgebra.jl Documentation

This is the initial documentation version.

GeometricAlgebra.AbstractMultivector — Type

AbstractMultivector{sig,C}

Supertype of all elements in the geometric algebra over a vector space with metric signature sig.

The parameter C is the type which the multivector components are stored as. For instance, a::AbstractMultivector{(1,1), SparseVector{Int64}} is a (mixed) multivector in the Euclidean plane with components of eltype(a) == Int64 stored in a Vector.

GeometricAlgebra.Blade — Type

Blade{sig,k,T,B} <: AbstractMultivector{sig,Pair{B,T}}
Blade{sig}(coeff, ublade)

A grade k blade (i.e., wedge product of k basis vectors) with coefficient of type T and unit blade represented by type B, in vector space of metric signature sig.

Unit blade type B	E.g. for $v_1∧v_3∧v_4$	Signature

Unsigned	0b1101	any

Vector{<:Integer}	[1, 3, 4]	any

Vector{<:Symbol}	[:v₁, :v₃, :v₄]	labelled

Examples

julia> Blade{(1,1,1)}(42, 0b101)
2-grade Blade{⟨+++⟩, Int64, UInt8, 2}:
 42 v₁v₃

julia> Blade{(x=1,y=1,z=1)}(1, [:x])
1-grade Blade{⟨x+,y+,z+⟩, Int64, Array{Symbol,1}, 1}:
 1 x

GeometricAlgebra.FixedGradeBlades — Type

FixedGradeBlades{B}(k, n)

Generates all grade k unit blades in n dimensions ordered lexicographically. Unit blades of type B are produced.

Examples

julia> FixedGradeBlades{UInt8}(2, 3) .|> bitstring
3-element Array{String,1}:
 "00000011"
 "00000101"
 "00000110"

julia> FixedGradeBlades{Vector{Int}}(3, 4) |> collect
4-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [1, 2, 4]
 [1, 3, 4]
 [2, 3, 4]

GeometricAlgebra.MixedMultivector — Type

MixedMultivector{sig,C} <: AbstractMultivector{sig,C}
MixedMultivector{sig}(comps)

Inhomogeneous multivector in vector space of metric signature sig, containing 2^dimension(sig) independent components of mixed grade.

Components are stored in the field .comps of type C, which in general may be an AbstractVector{T} or an AbstractDict{B,T}, where T is the component type (similar to Multivectors{sig,k,C}).

When C<:AbstractVector, components are ordered by the binary value of the unit blade. E.g., the coefficient of $v_1∧v_3∧v_4$ is stored at index 1 + Int(0b1101) == 14.

Examples

julia> MixedMultivector{(1,1,1)}(Dict(Int[] => 1, [1] => 2, [1,2] => 3))
MixedMultivector{⟨+++⟩, Dict{Array{Int64,1},Int64}}:
 1
 2 v₁
 3 v₁v₂

julia> basis((x=1, y=1), 1) + 7
MixedMultivector{⟨x+,y+⟩, Array{Float64,1}}:
 7.0
 1.0 x

julia> ans.comps
4-element Array{Float64,1}:
 7.0
 1.0
 0.0
 0.0

GeometricAlgebra.Multivector — Type

Multivector{sig,k,C} <: AbstractMultivector{sig,C}

Homogeneous multivector of grade k in space of metric signature sig, containing binomial(dimension(sig), k) independent components.

Components are stored in the field .comps of type C, which in general may be an AbstractVector{T} or an AbstractDict{B,T}, where T is the component type.

If C<:AbstractVector, components are stored in lexicographic order. E.g., the component of a 3-multivector corresponding to $v_1∧v_3∧v_4$ or 0b1101 is located at the ublade2lindex(0b1101) == 3rd index.

Base.eltype — Function

eltype(::AbstractMultivector)
eltype(::Type{<:AbstractMultivector})

Give the numerical type of the components of a multivector.

Base.keytype — Function

keytype(::AbstractMultivector)
keytype(::Type{<:AbstractMultivector})

Give the type in which unit blades are encoded in a multivector.

E.g., for Blade{sig,k,T,B}, this is B. For multivectors with components stored in a dictionary of type C, this is keytype(C). For components stored in a vector (where the unit blades are not stored explicitly), keytype(a) defaults to an unsigned type used to index the vector.

Examples

julia> Blade{(1,1,1)}(42, 0b101)
2-grade Blade{⟨+++⟩, Int64, UInt8, 2}:
 42 v₁v₃

julia> keytype(ans)
UInt8

julia> MixedMultivector{EuclideanSignature}(Dict([:z] => 1, [:x, :y] => 2))
MixedMultivector{EuclideanSignature, Dict{Array{Symbol,1},Int64}}:
 1 z
 2 xy

julia> keytype(ans)
Array{Symbol,1}

GeometricAlgebra.basis — Function

basis(::Type{<:AbstractMultivector}, i)
basis(::Type{<:AbstractMultivector})

Return the ith basis element of the given type, or a collection of all such basis elements if no index is given. For homogeneous types with grade k, this gives the ith basis k-blade in lexicographic order. For mixed multivector types, this gives the basis vectors / 1-blades.

Examples

julia> x, y, z = basis(Blade{(1,1,1),Float64,UInt,1})
3-element Array{Blade{⟨+++⟩, Float64, UInt64, 1},1}:
 1.0 v₁
 1.0 v₂
 1.0 v₃

julia> basis(Blade{(1,1,1),Float64,UInt,1}, 2) == y
true

julia> basis(Multivector{(1,1,1),Vector{Int},2}, 1) == x*y
true

GeometricAlgebra.basis — Method

basis(sig, i)
basis(sig)

Return a basis vector i for the space of metric signature sig, or a collection of all basis vectors (if sig has specified dimension).

Examples

julia> basis((x=1, y=1))
2-element Array{Blade{⟨x+,y+⟩, Float64, UInt8, 1},1}:
 1.0 x
 1.0 y

julia> basis((-1,1,1,1), 2)
1-grade Blade{⟨-+++⟩, Float64, UInt8, 1}:
 1.0 v₂

julia> basis(EuclideanSignature, :t)
1-grade Blade{EuclideanSignature, Float64, Array{Symbol,1}, 1}:
 1.0 t

GeometricAlgebra.bladeprodsign — Method

Compute sign flips of blade product due to transposing basis vectors. (The full sign of the product will also depend on the basis norms.)

GeometricAlgebra.blades — Method

blades(a::AbstractMultivector)

Return a vector of Blades representing the non-zero components of the multivector a.

Examples

julia> x, y, z = basis((1,1,1));

julia> blades(1 + 2x + 3y*z)
3-element Array{Blade{(1, 1, 1),Float64,UInt8,k} where k,1}:
 1.0
 2.0 v₁
 3.0 v₂v₃

GeometricAlgebra.commutator_prod — Method

a×b
commutator_prod(a, b)

Commutator product of multivectors, given by $a × b = (a*b - b*a)/2$.

GeometricAlgebra.dimension — Method

dimension(::AbstractMultivector)
dimension(::Type{<:AbstractMultivector})

Return the dimension of the vector space underlying the given multivector (or multivector type).

GeometricAlgebra.dot — Method

a⋅b
dot(a, b)

"Fat" dot product of multivectors, giving the lowest-possible-grade part of the geometric product. Defined by $a⋅b = ∑_{ij} ⟨⟨a⟩_i * ⟨b⟩_j⟩_|i - j|$.

GeometricAlgebra.grade — Method

grade(a::AbstractMultivector)

The grade of the homogeneous multivector a. For MixedMultivectors, see grades.

Examples

julia> x, y, z = basis((1,1,1))
3-element Array{Blade{⟨+++⟩, Float64, UInt8, 1},1}:
 1.0 v₁
 1.0 v₂
 1.0 v₃

julia> grade(x*y)
2

GeometricAlgebra.grade — Method

grade(a::AbstractMultivector, k)

The grade k part of a (mixed) multivector a. Returns a grade k Multivector.

Examples

julia> grade(x + 1 + y + x*y, 1)
1-grade Multivector{⟨x+,y+⟩, Array{Float64,1}, 1}:
 1.0 x
 1.0 y

GeometricAlgebra.grades — Method

grades(a::AbstractMultivector)

The non-zero grades of the multivector a.

Examples

julia> x, y, z = basis((1,1,1))
3-element Array{Blade{⟨+++⟩, Float64, UInt8, 1},1}:
 1.0 v₁
 1.0 v₂
 1.0 v₃

julia> grades(1 + x*y)
2-element Array{Int64,1}:
 0
 2

GeometricAlgebra.hodgedual — Method

★(a)
hodgedual(a)

Hodge dual or 'star' operator, satisfying $a∧★(b) == (a∗b)vol(a)$.(?)

GeometricAlgebra.lindex2ublade — Method

Convert linear index i of into unit k-blade. The linear index (lindex) gives the index of the unit blade sorted lexicographically.

julia> lindex2ublade.(UInt8, 3, 1:5) .|> bitstring
5-element Vector{String}:
 "00000111"
 "00001011"
 "00001101"
 "00001110"
 "00010011"

GeometricAlgebra.mapcomps — Function

mapcomps!(f, a)
mapcomps(f, a)

Map the components of a multivector a by the function f, which is called with a single Blade argument and whose return type should be eltype(a).

The mutating version mapcomps! maps the components in-place.

GeometricAlgebra.mapcomps! — Function

mapcomps!(f, a)
mapcomps(f, a)

Map the components of a multivector a by the function f, which is called with a single Blade argument and whose return type should be eltype(a).

The mutating version mapcomps! maps the components in-place.

GeometricAlgebra.next_bit_permutation — Method

Return the smallest uint larger than the one given which has the same number of binary ones. Algorithm is Gosper's hack, documented at http://graphics.stanford.edu/~seander/bithacks.html#NextBitPermutation.

julia> GeometricAlgebra.next_bit_permutation(0b1011) |> bitstring
"00001101"

GeometricAlgebra.next_ublade — Method

Give the next unit blade of the same grade in lexicographic order.

GeometricAlgebra.rsqrt — Method

rsqrt(x)

Real-valued square root defined on all real numbers as sign(x)sqrt(abs(x)).

Examples

julia> rsqrt(-9)
-3.0

GeometricAlgebra.show_blade — Method

Display blade with parentheses surrounding coefficient if necessary.

GeometricAlgebra.show_mixedmultivector — Method

Display inhomogeneous MixedMultivector with each grade on a new line.

GeometricAlgebra.show_multivector — Method

Display a multivector as a column of blades, with coefficients aligned using the native alignment mechanism, and blades basis aligned.

julia> 1e3x + y + 1e-3z
1-Multivector{⟨x+,y+,z+⟩, Vector{Float64}, 1}
 1000.0   x
    1.0   y
    0.001 z

GeometricAlgebra.show_ublade — Method

Display unit blade without coefficient.

GeometricAlgebra.signature — Method

signature(::AbstractMultivector)
signature(::Type{<:AbstractMultivector})

Return the metric signature associated with the given multivector or multivector type.

Metric signatures can be any object implementing getindex to return the norm of a given basis vector. Signatures may be Tuples, NamedTuples, or instances of AbstractMetricSignature.

Examples

julia> x = Blade{(x=1, y=1)}(1, [:x])
1-grade Blade{⟨x+,y+⟩, Int64, Array{Symbol,1}, 1}:
 1 x

julia> signature(x)
(x = 1, y = 1)

GeometricAlgebra.ubladeprod — Method

Compute the geometric product of unit basis vectors, returning a scalar coefficient and the resulting unit blade.

julia> GeometricAlgebra.ubladeprod((1,1,1), [1, 2, 1, 2, 3])
(-1, [3])

julia> GeometricAlgebra.ubladeprod((1,1,1), 0b011, 0b111)
(-1, 0x04)

GeometricAlgebra.vol — Method

vol(x)

The volume form or psuedoscalar element.

Examples

julia> x = basis((1, 1, 1), 1)
1-grade Blade{⟨+++⟩, Float64, UInt8, 1}:
 1.0 v₁

julia> vol(x)
3-grade Blade{⟨+++⟩, Float64, UInt8, 3}:
 1.0 v₁v₂v₃

GeometricAlgebra.wedge — Method

a∧b
wedge(a, b)

Wedge product of multivectors, giving the highest-possible-grade part of the geometric product a*b. Defined by $a∧b = ∑_{ij} ⟨⟨a⟩_i * ⟨b⟩_j⟩_(i + j)$.