GeometricAlgebra.jl

Documentation for GeometricAlgebra.jl

GeometricAlgebra.AbstractMultivector — Type

AbstractMultivector{sig}

Supertype of all elements in the geometric algebra over the vector space with metric signature sig.

GeometricAlgebra.Blade — Type

Blade{sig,k,bits,T} <: HomogeneousMultivector{sig,k}

A blade of grade k with basis blade bits and scalar coefficient of type T.

Parameters

sig: metric signature defining the parent geometric algebra

k: grade of the blade, equal to grade(bits) === count_ones(bits)

bits: unsigned integer representing the indices of the basis vectors whose wedge product is the unit blade

T: type of the scalar coefficient

GeometricAlgebra.HomogeneousMultivector — Type

HomogeneousMultivector{sig,k} <: AbstractMultivector{sig}

Supertype of grade k elements in the geometric algebra with metric signature sig.

GeometricAlgebra.MixedMultivector — Type

MixedMultivector{sig,S} <: AbstractMultivector{sig}

An generally inhomogeneous multivector.

All elements of a geometric algebra are representable as a MixedMultivector.

Parameters

sig: metric signature defining the parent geometric algebra

S: type in which the multivector components are stored; usually a vector-like or dictionary-like type

GeometricAlgebra.Multivector — Type

Multivector{sig,k,S} <: HomogeneousMultivector{sig,k}

A homogeneous multivector of grade k with storage type S.

Parameters

sig: metric signature defining the parent geometric algebra

k: grade of the multivector

S: type in which the multivector components are stored; usually a vector-like or dictionary-like type

Base.:* — Method

*(::AbstractMultivector, ::AbstractMultivector)

Geometric product. See GeometricAlgebra.geometric_prod.

Base.:~ — Function

~(a::AbstractMultivector)
reversion(a::AbstractMultivector)

The reverse of a multivector.

The reversion antiautomorphism $ρ$ is defined by $ρ(u) = u$ on vectors and $ρ(ab) = ρ(b)ρ(a)$ on higher-grade elements.

Base.eltype — Method

eltype(a)

The numerical type of the components of a multivector or multivector type.

GeometricAlgebra.bits_to_indices — Method

bits_to_indices(bits)

Return the positions of the ones in the unsigned integer bits.

Used to convert between representations of a unit blade. See also indices_to_bits.

Examples

julia> GeometricAlgebra.bits_to_indices(0b1001101)
4-element Vector{Int64}:
 1
 3
 4
 7

GeometricAlgebra.factor_from_squares — Method

Compute the overall factor arising from the geometric product between repeated basis vectors.

GeometricAlgebra.geometric_prod — Function

a*b
geometric_prod(a, b)

Geometric product of multivectors.

If both arguments are blades, then the result is a Blade. Otherwise, the geometric product is general MixedMultivector.

GeometricAlgebra.graded_prod — Function

graded_prod(a, b, grade_selector::Function)

Homogenised geometric product with resulting grade given as a function of the grades of a and b.

If $a_p$ and $b_q$ are multivectors of grade $p$ and $q$ respectively, then this is equal to $⟨a_p b_q⟩_{k(p,q)}$ where $k(p,q)$ is given by grade_selector. If a or b are multi-graded, then this extends linearly to $∑_{pq} ⟨⟨a⟩_p ⟨b⟩_q⟩_{k(p,q)}$.

GeometricAlgebra.homogeneous_prod — Function

homogeneous_prod(a, b, Val(k))

The grade k part of the geometric product of a and b. Equivalent to, but more efficient than, grade(geometric_prod(a, b), k).

GeometricAlgebra.indices_to_bits — Method

indices_to_bits(indices)

Create unsigned integer with bits at the positions given in the vector indices.

Used to convert between representations of a unit blade. See also bits_to_indices.

Examples

julia> GeometricAlgebra.indices_to_bits([1, 2, 5]) |> UInt16 |> bitstring
"0000000000010011"

GeometricAlgebra.involute — Method

involute(a::AbstractMultivector)

The involute of a multivector, negating the odd-grade part.

The involution automorphism $ι$ is defined by $ι(u) = -u$ on vectors and $ι(ab) = ι(a)ι(b)$ on higher-grade elements.

GeometricAlgebra.next_bit_permutation — Method

Return the smallest uint larger than the one given which has the same number of binary ones. Algorithm is Gosper's hack.

julia> GeometricAlgebra.next_bit_permutation(0b1011) |> bitstring
"00001101"

GeometricAlgebra.reversion — Function

~(a::AbstractMultivector)
reversion(a::AbstractMultivector)

The reverse of a multivector.

The reversion antiautomorphism $ρ$ is defined by $ρ(u) = u$ on vectors and $ρ(ab) = ρ(b)ρ(a)$ on higher-grade elements.

GeometricAlgebra.show_blade — Method

Display blade with parentheses surrounding coefficient if necessary.

GeometricAlgebra.show_mixedmultivector — Method

Display inhomogeneous MixedMultivector with each grade on a new line.

GeometricAlgebra.show_multivector — Method

Display a multivector as a column of blades, with coefficients aligned using the native alignment mechanism, and blades basis aligned.

julia> 1e3x + y + 1e-3z
1-Multivector{⟨x+,y+,z+⟩, Vector{Float64}, 1}
 1000.0   x
    1.0   y
    0.001 z

GeometricAlgebra.show_ublade — Method

Display unit blade without coefficient.

GeometricAlgebra.sign_from_swaps — Method

Compute sign flips of blade product due to transposing basis vectors. (The full sign of the product will also depend on the basis norms.)