API

GeometricAlgebra.Blade

GeometricAlgebra.Multivector

GeometricAlgebra.UnitBlade

GeometricAlgebra.:∧

GeometricAlgebra.:⋅

GeometricAlgebra.grade_els

GeometricAlgebra.grade_index

GeometricAlgebra.is_homogeneous

GeometricAlgebra.lcontract

GeometricAlgebra.rcontract

GeometricAlgebra.@basis

GeometricAlgebra.Blade — Type

Blade{B,T}

Blade living in the subspace spanned by B, with a coefficient of type T. Can be interpreted as a scaled version of a UnitBlade.

GeometricAlgebra.Multivector — Type

Linear combination of blades, forming a general multivector.

GeometricAlgebra.UnitBlade — Type

UnitBlade{S,G,I}

Unit blade with grade G, indices I, and Signature S.

GeometricAlgebra.:∧ — Function

`x ∧ y`

Outer product of x with y.

GeometricAlgebra.:⋅ — Function

`x ⋅ y`

Inner product of x with y.

GeometricAlgebra.grade_els — Method

Return the blades of grade g from a Multivector.

GeometricAlgebra.grade_index — Method

`grade_index(dim, i)`

Return the grade index of i.

Example

julia> grade_index(3, [1])
1

julia> grade_index(3, [1, 2])
1

julia> grade_index(3, [3, 1])
3

GeometricAlgebra.is_homogeneous — Method

Whether the mv only contains elements of a single grade.

GeometricAlgebra.lcontract — Function

`lcontract(x, y)`

Left contraction of x with y.

GeometricAlgebra.rcontract — Function

`rcontract(x, y)`

Right contraction of x with y.

GeometricAlgebra.@basis — Macro

@basis [mod=Main, prefix=:v, signature="Ø"] <dim>

Pull in all unit blade symbols from a dim-dimensional geometric algebra with a given signature. The symbols are evaluated inside the module mod, prefixed with prefix.