Reference

GeometricAlgebra.GeometricAlgebra — Module

GeometricAlgebra

Implements multivector (or $k$-vector) types from geometric algebra (a.k.a. Clifford algebra).

Exported Types

                   AbstractMultivector{Sig}
                     /                  \
   HomogeneousMultivector{Sig,K}    MixedMultivector{Sig,S}
       /                \                             
Blade{Sig,K,T}    Multivector{Sig,K,S}                
                                                   
                  ╰───── CompositeMultivector{Sig,S} ─────╯

See basis and @basis to get started.

GeometricAlgebra.AbstractMultivector — Type

AbstractMultivector{Sig}

Supertype of all elements in the geometric algebra defined by the metric signature Sig.

Subtypes

                   AbstractMultivector{Sig}
                     /                  \
   HomogeneousMultivector{Sig,K}    MixedMultivector{Sig,S}
       /                \                             
Blade{Sig,K,T}    Multivector{Sig,K,S}                
                                                   
                  ╰───── CompositeMultivector{Sig,S} ─────╯

Blade: a scalar multiple of a wedge product of orthogonal basis vectors.

Multivector: a homogeneous multivector; a sum of same-grade blades.

MixedMultivector: an inhomogeneous multivector. All elements in a geometric algebra can be represented as this type (though not most efficiently).

Note

The mathematical definition of a $k$-blade is the wedge product of $k$ vectors, not necessarily basis vectors. Thus, not all $k$-blades are representable as a Blade, but are always representable as a sum of Blades, or as a Multivector.

Type Parameters

Sig: The metric signature which defines the geometric algebra. This can be any all-bits value which satisfies the metric signature interface. For example, (1, 1, 1) or EuclideanSignature(3) both define the standard geometric algebra of $ℝ^3$.

T: The numerical type of the coefficient of a Blade.

K: An Int specifying the grade of a HomogeneousMultivector.

S: The storage type of the components of a CompositeMultivector. This is assumed to be mutable, and is usually a subtype of Vector, MVector or SparseVector.

GeometricAlgebra.BitPermutations — Type

BitPermutations{T}(n)

Infinite iterator returning all unsigned integers of type T, in ascending order, for which Base.count_ones is n.

GeometricAlgebra.Blade — Type

Blade{Sig,K,T} <: HomogeneousMultivector{Sig,K}

A blade of grade K with basis blade bits and scalar coefficient of type T.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K: Grade of the blade, equal to count_ones(bits), retrieved with grade().

T: Numerical type of the scalar coefficient, retrieved with eltype().

GeometricAlgebra.Blade — Method

Blade{Sig}(bits, coeff)
Blade{Sig}(bits => coeff)

Basis blade with indices encoded by bits and scalar coefficient coeff.

Examples

julia> Blade{3}(0b110 => 42) # a grade 2 blade in 3 dimensions
Blade{3, 2, Int64}:
 42 v23

GeometricAlgebra.Cl — Type

Cl(p, q=0, r=0)

Metric signature where p, q and r is the number of basis vectors of norm +1, -1 and 0, respectively.

Examples

julia> basis(Cl(1,3))
4-element Vector{Blade{Cl(1,3), 1, Int64}}:
 1v1
 1v2
 1v3
 1v4

julia> ans .^ 2
4-element Vector{Blade{Cl(1,3), 0, Int64}}:
  1
 -1
 -1
 -1

GeometricAlgebra.HomogeneousMultivector — Type

HomogeneousMultivector{Sig,K} <: AbstractMultivector{Sig}

Abstract supertype of Blade and Multivector.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K: Grade of the blade or multivector, retrieved with grade().

GeometricAlgebra.MixedMultivector — Type

MixedMultivector{Sig,S} <: AbstractMultivector{Sig}

A (possibly inhomogeneous) multivector.

All elements of a geometric algebra are representable as a MixedMultivector.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

S: Storage type of the multivector components, usually a subtype of AbstractVector.

GeometricAlgebra.MixedMultivector — Method

MixedMultivector{Sig}(comps)

Inhomogeneous multivector with components vector comps. The components are ordered first by grade then lexicographically (see GeometricAlgebra.mmv_bits).

Examples

julia> MixedMultivector{3}(1:2^3)
8-component MixedMultivector{3, UnitRange{Int64}}:
 1
 2 v1 + 3 v2 + 4 v3
 5 v12 + 6 v13 + 7 v23
 8 v123

julia> grade(ans, 1)
3-component Multivector{3, 1, UnitRange{Int64}}:
 2 v1
 3 v2
 4 v3

GeometricAlgebra.Multivector — Type

Multivector{Sig,K,S} <: HomogeneousMultivector{Sig,K}

A homogeneous multivector of grade K with storage type S.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K: Grade of the multivector, retrieved with grade().

S: Storage type of the multivector components, usually a subtype of AbstractVector.

GeometricAlgebra.Multivector — Method

Multivector{Sig,K}(comps)

Multivector of grade K with components vector comps.

Examples

julia> Multivector{3,2}(1:3) # 3D bivector
3-component Multivector{3, 2, UnitRange{Int64}}:
 1 v12
 2 v13
 3 v23

Base.eltype — Method

eltype(::AbstractMultivector)

The numerical type of the components of a multivector instance or type.

GeometricAlgebra.basis — Method

basis(sig; grade=1)

Return basis blades for the geometric algebra defined by the metric signature sig.

Examples

julia> basis(3)
3-element Vector{Blade{3, 1, Int64}}:
 1v1
 1v2
 1v3

julia> basis(Cl(1,3); grade=2)
6-element Vector{Blade{Cl(1,3), 2, Int64}}:
 1v12
 1v13
 1v23
 1v14
 1v24
 1v34

GeometricAlgebra.bits_of_grade — Method

bits_of_grade(k[, dim])

Generate basis blade bits of grade k in ascending order. Yields all basis blades in the dimension dim, if given, otherwise iterate indefinitely.

Examples

julia> GeometricAlgebra.bits_of_grade(2, 4) .|> UInt8 .|> bitstring
6-element Vector{String}:
 "00000011"
 "00000101"
 "00000110"
 "00001001"
 "00001010"
 "00001100"

GeometricAlgebra.bits_to_indices — Method

bits_to_indices(bits)

Return the positions of the ones in the unsigned integer bits.

Used to convert between representations of a unit blade. Inverse of indices_to_bits.

Examples

julia> GeometricAlgebra.bits_to_indices(0b1001101)
4-element Vector{Int64}:
 1
 3
 4
 7

GeometricAlgebra.bits_to_mmv_index — Method

bits_to_mmv_index(bits::Unsigned)

Convert a unit blade bits to a linear index for accessing components of a MixedMultivector.

GeometricAlgebra.bits_to_mv_index — Method

bits_to_mv_index(bits::Unsigned)

Convert a unit blade bits to a linear index for accessing components of a Multivector.

GeometricAlgebra.componentstype — Method

componentstype(sig, N, T)

Array type to use the components for multivectors of signature sig. The resulting type should be able to store N components (in the case of a fixed-size array) of element type T.

This is used when converting a Blade into a CompositeMultivector to determine the type of the components array.

GeometricAlgebra.dimension — Method

dimension(::AbstractMultivector)

The dimension of the underlying vector space of the geometric algebra. See ncomponents for the dimension of the algebra (i.e., the number of independent components of a general multivector).

GeometricAlgebra.factor_from_squares — Method

factor_from_squares(sig, bits::Unsigned)

Compute the overall factor arising from the geometric product between repeated basis vectors.

GeometricAlgebra.generated_multivector_function — Method

generated_multivector_function(f, a, b, ...)

Trace evaluation of f(a, b, ...)::CompositeMultivector on symbolic versions of each AbstractMultivector instance or type a, b, ..., returning an expression body suitable for a @generated function.

The names of the arguments to be passed to the function body are the literal symbols :a, :b, etc.

Examples

julia> u, v = Multivector.(basis(2))
2-element Vector{Multivector{2, 1, Vector{Int64}}}:
 1v1
 1v2

julia> using MacroTools: prettify

julia> ex = GeometricAlgebra.generated_multivector_function(*, u, v) |> prettify
:(let a = (CompositeMultivector(a)).components, b = (CompositeMultivector(b)).components
      comps = create_array(Vector{Any}, Int64, Val{1}(), Val{(4,)}(), a[1] * b[1] + a[2] * b[2], 0, 0, a[1] * b[2] + (-1 * a[2]) * b[1])
      (MixedMultivector{2})(comps)
  end)

GeometricAlgebra.geometric_prod_bits — Method

geometric_prod_bits(sig, a::Unsigned, b::Unsigned)

Compute the geometric product between unit blades. Returns a tuple of the overall scalar factor and the resulting unit blade.

GeometricAlgebra.grade — Method

grade(::HomogeneousMultivector{Sig,K}) -> K

The grade of a homogeneous multivector (a Blade or Multivector) instance or type.

GeometricAlgebra.indices_to_bits — Method

indices_to_bits(indices)

Create unsigned integer with bits at the positions given in the vector indices.

Used to convert between representations of a unit blade. Inverse of bits_to_indices.

Warning

Produces incorrect results if elements of indices are greater than the number of bits in bits_scalar() <: Unsigned.

Examples

julia> GeometricAlgebra.indices_to_bits([1, 2, 5]) |> UInt16 |> bitstring
"0000000000010011"

GeometricAlgebra.mmv_bits — Method

mmv_bits(::Val{N})

Vector of unit blades corresponding to components of an N-dimensional MixedMultivector. Mixed multivector components are ordered first by grade then by numerical value of the unit blade, so that the grade K components are contiguous and given by mv_bits(Val(N), Val(K))

Examples

julia> GeometricAlgebra.mmv_bits(Val(3)) .|> UInt8 .|> bitstring
8-element Vector{String}:
 "00000000"
 "00000001"
 "00000010"
 "00000100"
 "00000011"
 "00000101"
 "00000110"
 "00000111"

GeometricAlgebra.mv_bits — Method

mv_bits(::Val{N}, ::Val{K})

Vector of unit blades corresponding to components of an N-dimensional Multivector of grade K. Multivector components are sorted by the numerical value of the unit blade.

Examples

julia> GeometricAlgebra.mv_bits(Val(4), Val(2)) .|> UInt8 .|> bitstring
6-element Vector{String}:
 "00000011"
 "00000101"
 "00000110"
 "00001001"
 "00001010"
 "00001100"

GeometricAlgebra.ncomponents — Method

ncomponents(sig)
ncomponents(sig, k)

Dimension of (the grade-k subspace of) the geometric algebra of metric signature sig, viewed as a vector space.

If the dimension of the underlying vector space in $n$, then the algebra is $2^n$-dimensional, and its grade-$k$ subspace $\binom{n}{k}$-dimensional.

GeometricAlgebra.ncomponents — Method

ncomponents(::CompositeMultivector)

Number of independent components of a multivector instance or type.

In $n$ dimensions, this is $\binom{n}{k}$ for a Multivector and $2^n$ for a MixedMultivector.

GeometricAlgebra.next_bit_permutation — Method

Return the smallest uint larger than the one given which has the same number of binary ones. Algorithm is Gosper’s hack.

Examples

julia> GeometricAlgebra.next_bit_permutation(0b1011) |> bitstring
"00001101"

GeometricAlgebra.show_basis_blade — Method

show_basis_blade(io, sig, indices)

Show the basis blade which is the product of the unit vectors in indices in a geometric algebra defined by sig. Methods dispatching on sig should be added to customise basis blade labels for particular algebras.

The fallback method is:

show_basis_blade(io, sig, indices) = printstyled(io, "v"*join(string.(indices)); bold=true)

Examples

julia> GeometricAlgebra.show_basis_blade(stdout, (1, 1, 1), [1, 3])
v13

julia> using GeometricAlgebra: subscript

julia> GeometricAlgebra.show_basis_blade(io, sig, indices) = print(io, join("𝒆".*subscript.(indices), "∧"))

julia> prod(basis(4))
Blade{⟨++++⟩, 4, Int64} of grade 4:
 1 𝒆₁∧𝒆₂∧𝒆₃∧𝒆₄

GeometricAlgebra.show_blade — Method

Display blade with parentheses surrounding coefficient if necessary.

julia> GeometricAlgebra.show_blade(stdout, Blade{(x=1,)}(0b1 => 1 + im))
(1+1im) x

GeometricAlgebra.show_mixedmultivector — Method

Display an inhomogeneous MixedMultivector with each grade on a new line.

GeometricAlgebra.show_multivector — Method

Display homogeneous multivector components as a column of blades, with coefficients and blades aligned using the native alignment mechanism.

julia> a = Multivector{(1,1,1),1}([1e3, 1, 1e-3]);

julia> GeometricAlgebra.show_multivector(stdout, a)
1000.0   v1
   1.0   v2
   0.001 v3

GeometricAlgebra.show_signature — Method

show_signature(io, sig)

Pretty-print the metric signature sig.

This is used to display the metric signature type parameter in AbstractMultivector subtypes to reduce visual noise. Methods may optionally be added for user-defined metric signatures, in a similar fashion to Base.show.

Examples

julia> sig = (+1,-1,-1,-1)
(1, -1, -1, -1)

julia> GeometricAlgebra.show_signature(stdout, sig)
⟨+---⟩

julia> Blade{sig}
Blade{⟨+---⟩} (pretty-printed Blade{(1, -1, -1, -1)})

GeometricAlgebra.sign_from_swaps — Method

Compute sign flips of blade product due to transposing basis vectors into sorted order. (The full sign of the product will also depend on the basis norms.)

GeometricAlgebra.signature — Method

signature(::AbstractMultivector{Sig}) -> Sig

The metric signature type parameter of the multivector instance or type.

GeometricAlgebra.@basis — Macro

@basis sig

Populate namespace with basis blades of every grade in the geometric algebra with metric signature sig.

See also @basisall.

Examples

julia> @basis 3
[ Info: Defined basis blades v, v1, v2, v3, v12, v13, v23, v123

julia> 1v2 + 3v12
8-component MixedMultivector{3, Vector{Int64}}:
 1 v2
 3 v12

GeometricAlgebra.@basisall — Macro

@basisall sig

Similarly to @basis, populate namespace with basis blades, but include all permutations of each blade.

Warning

This defines $2^n$ variables for an $n$ dimensional signature!

Examples

julia> @basisall (+1,-1)
[ Info: Defined basis blades v, v1, v2, v12, v21

julia> v12 == -v21
true