Reference

Multivector Types

GeometricAlgebra.AbstractMultivector — Type

AbstractMultivector{Sig}

Supertype of all elements in the geometric algebra defined by the metric signature Sig.

Subtypes

         AbstractMultivector{Sig}
            /               \                             
BasisBlade{Sig,K,T}   Multivector{Sig,K,S}                

BasisBlade: a scalar multiple of a wedge product of orthogonal basis vectors.

Multivector: a homogeneous or inhomogeneous multivector; a sum of basis blades.

GeometricAlgebra.BasisBlade — Type

BasisBlade{Sig,K,T}

A basis blade of grade K and scalar coefficient of type T.

Basis blades are scalar multiples of wedge products of orthogonal basis vectors.

Note

Not every $k$-blade (i.e., wedge product of $k$ linearly independent vectors) is representable as a BasisBlade. However, every $k$-blade is a Multivector of grade k.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K::Int: Grade of the blade, equal to count_ones(bits), retrieved with grade().

T: Numerical type of the scalar coefficient, retrieved with eltype().

GeometricAlgebra.BasisBlade — Method

BasisBlade{Sig}(bits, coeff)
BasisBlade{Sig}(bits => coeff)

Basis blade with indices encoded by bits and scalar coefficient coeff.

Indices are encoded in binary (e.g., $v₁∧v₃∧v₄$ has bits 0b1101).

Examples

julia> BasisBlade{3}(0b110 => 42) # a grade 2 blade in 3 dimensions
BasisBlade{3, 2, Int64}:
 42 v23

GeometricAlgebra.Multivector — Type

Multivector{Sig,K,S} <: AbstractMultivector{Sig}

A general multivector with parts of grade ∈ K.

For homogeneous k-vectors, the grade parameter K is an integer. Inhomogeneous multivectors may be specified with a range or tuple of grades.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K: Grade(s) present in the multivector. Can be an integer or a collection of integers (a range or tuple).

S: Storage type of the multivector components, usually a subtype of AbstractVector.

GeometricAlgebra.Multivector — Method

Multivector{Sig,K}(comps)

Multivector with grade(s) K and component vector comps.

Components are ordered first by grade, then lexicographically by bits (see componentbits).

Examples

julia> Multivector{3,0:3}(1:2^3)
8-component Multivector{3, 0:3, UnitRange{Int64}}:
 1
 2 v1 + 3 v2 + 4 v3
 5 v12 + 6 v13 + 7 v23
 8 v123

julia> grade(ans, 1)
3-component Multivector{3, 1, UnitRange{Int64}}:
 2 v1
 3 v2
 4 v3

Base.eltype — Method

eltype(::AbstractMultivector)

The numerical type of the components of a multivector instance (or type).

GeometricAlgebra.componentindex — Method

componentindex(a::Multivector, b)

Index of components vector of a corresponding to the BasisBlade or bits b.

GeometricAlgebra.dimension — Method

dimension(sig)
dimension(::AbstractMultivector)

The dimension of the underlying vector space of the geometric algebra. See ncomponents for the dimension of the algebra (i.e., the number of independent components of a general multivector).

GeometricAlgebra.grade — Method

grade(a)

Grade(s) of a multivector a.

The grade of a BasisBlade{Sig,K} or Multivector{Sig,K} is the second type parameter, K, which may be an integer (if a is homogeneous) or a collection (a range or tuple of grades).

See also ishomogeneous.

GeometricAlgebra.ncomponents — Method

ncomponents(::CompositeMultivector)

Number of independent components of a multivector instance (or type).

GeometricAlgebra.resulting_multivector_type — Method

resulting_multivector_type(f, a, b, ...)

Return a Multivector{Sig,K,S} type with parameters (signature S, grade(s) K and storage type S) appropriate for representing the result of f(a, b).

GeometricAlgebra.signature — Method

signature(::AbstractMultivector{Sig}) = Sig

The metric signature type parameter of the multivector instance (or type).

Algebraic Operations

Base.:~ — Method

~a
reversion(a::AbstractMultivector)

Reversion of a multivector.

Reversion is an anti-automorphism defined by reversing the order of the geometric product: ~(a*b) == ~b * ~a. For a k-vector a, the reversion is reversion_sign(k)*a where the sign is given by $(-1)^{k(k - 1)/2}$.

See also involution and clifford_conj.

GeometricAlgebra.:'ᶜ — Method

a'ᶜ
clifford_conj(a)

Clifford conjugate of a multivector.

Equivalent to reversion(involution(a)).

GeometricAlgebra.:∧ — Method

a ∧ b
wedge(a, b)

Wedge product of multivectors, a.k.a. the outer, exterior or alternating product.

This is a grade-raising operation, equivalent to graded_prod(+, a, b). If a and b are of grades $p$ and $q$ respectively, then a ∧ b is the grade $p + q$ part of a*b.

GeometricAlgebra.:⊙ — Method

a ⊙ b
scalar_prod(a, b) -> Number

Scalar part of the multivector product a*b.

GeometricAlgebra.:⋅ — Method

a ⋅ b
inner(a, b)

Inner product of multivectors.

This is a grade lowering operation, equivalent to graded_prod(abs∘-, a, b). If a and b are of grades $p$ and $q$ respectively, then a ⋅ b is the grade $|p - q|$ part of a*b.

Note that for scalars a and b, the inner product reduces to scalar multiplication, in contrast to some authors (see [D02] for discussion).

See also lcontract and rcontract.

GeometricAlgebra.:⨼ — Method

a ⨼ b
lcontract(a, b)

Left contraction of multivectors. See also rcontract and inner.

Equivalent to graded_prod((p, q) -> q - p, a, b). If a and b are of grades $p$ and $q$ respectively, then a ⨼ b is the grade $q - p$ part of a*b.

GeometricAlgebra.:⨽ — Method

a ⨽ b
rcontract(a, b)

Left contraction of multivectors. See also lcontract and inner.

Equivalent to graded_prod(-, a, b). If a and b are of grades $p$ and $q$ respectively, then a ⨽ b is the grade $p - q$ part of a*b.

GeometricAlgebra.clifford_conj — Method

a'ᶜ
clifford_conj(a)

Clifford conjugate of a multivector.

Equivalent to reversion(involution(a)).

GeometricAlgebra.flipdual — Function

flipdual(a)

A dual of a multivector, for when the overall sign isn’t important.

For a unit basis blade a::BasisBlade, the flipdual satisfies a*flipdual(a) == ±I where ±I is the unit pseudoscalar or its negative.

Computing the flipdual is cheap, and is its own inverse: for a BasisBlade, its bits are flipped, and for a CompositeMultivector, the components vector is simply reversed.

The flipdual is metric independent, but depends on a choice of basis. It differs from the Hodge and Poincaré duals by a per-grade scalar factor. This makes it useful in projective geometry, where scalar factors are largely arbitrary.

See also hodgedual and poincaredual.

GeometricAlgebra.geometric_prod — Method

a * b
geometric_prod(a, b)

Geometric product of multivectors.

GeometricAlgebra.graded_multiply — Method

graded_multiply(f, a::AbstractMultivector)

Multiply the grade k part of a by f(k).

GeometricAlgebra.graded_prod — Method

graded_prod(grade_selector::Function, a, b)

A "graded" product of multivectors, generalising the wedge $∧$, inner $⋅$ and contraction products. For example, the wedge product is defined as:

wedge(a, b) = graded_prod(+, a, b)

If grade(a) == p and grade(b) == q, then graded_prod(f, a, b) is the grade f(p, q) part of a*b. The definition extends linearly to general multivectors $A$ and $B$ as

\[ (A, B) ↦ \sum_{p,q} ⟨⟨A⟩_p ⟨B⟩_q⟩_{f(p, q)}\]

where $⟨⋅⟩_k$ denotes the grade $k$ part.

GeometricAlgebra.hodgedual — Function

hodgedual(a) = ~a*I

Hodge dual of a multivector.

The Hodge dual is given by

\[⋆a = ã I\]

where $ã$ is the reversion of $a$ and $I$ is the unit pseudoscalar. For $k$-vectors $a$ and $b$, it is alternatively defined by

\[a ∧ ⋆b = ⟨a, b⟩ I\]

where $⟨a, b⟩ = a ⊙ b̃$ is the induced inner product on $k$-vectors.

The Hodge dual depends on the metric and orientation (choice of pseudoscalar).

See also poincaredual.

Examples

julia> u = Multivector{3,1}(1:3)
3-component Multivector{3, 1, UnitRange{Int64}}:
 1 v1
 2 v2
 3 v3

julia> hodgedual(u)
3-component Multivector{3, 2, Vector{Int64}}:
  3 v12
 -2 v13
  1 v23

julia> u ∧ hodgedual(u), u ⊙ ~u
(14v123, 14)

GeometricAlgebra.inner — Method

a ⋅ b
inner(a, b)

Inner product of multivectors.

This is a grade lowering operation, equivalent to graded_prod(abs∘-, a, b). If a and b are of grades $p$ and $q$ respectively, then a ⋅ b is the grade $|p - q|$ part of a*b.

Note that for scalars a and b, the inner product reduces to scalar multiplication, in contrast to some authors (see [D02] for discussion).

See also lcontract and rcontract.

GeometricAlgebra.involution — Method

involution(a)

Involute of a multivector.

Involution is an automorphism defined by reflecting through the origin: for homogeneous multivectors, involution(a) == (-1)^grade(a)*a.

See also reversion and clifford_conj.

GeometricAlgebra.lcontract — Method

a ⨼ b
lcontract(a, b)

Left contraction of multivectors. See also rcontract and inner.

Equivalent to graded_prod((p, q) -> q - p, a, b). If a and b are of grades $p$ and $q$ respectively, then a ⨼ b is the grade $q - p$ part of a*b.

GeometricAlgebra.poincaredual — Function

poincaredual(a)

Poincaré dual of a multivector.

For a unit basis blade a::BasisBlade, the Poincaré dual satisfies a*poincaredual(a) == I where I is the unit pseudoscalar. I.e., poincaredual(a) is a “right complement” of a.

The Poincaré dual is metric independent, but depends on a choice of basis. This makes is useful in degenerate algebras: non-zero multivectors have non-zero Poincaré duals, even if their Hodge dual is zero.

See also hodgedual and flipdual.

Examples

julia> @basis Cl(2,0,1)
[ Info: Defined basis blades v, v1, v2, v3, v12, v13, v23, v123

julia> hodgedual(v3), v3*v123 # Hodge dual is zero because v3*v3 == 0
(0, 0)

julia> poincaredual(v3)
BasisBlade{Cl(2,0,1), 2, Int64}:
 1 v12

GeometricAlgebra.rcontract — Method

a ⨽ b
rcontract(a, b)

Left contraction of multivectors. See also lcontract and inner.

Equivalent to graded_prod(-, a, b). If a and b are of grades $p$ and $q$ respectively, then a ⨽ b is the grade $p - q$ part of a*b.

GeometricAlgebra.reversion — Method

~a
reversion(a::AbstractMultivector)

Reversion of a multivector.

Reversion is an anti-automorphism defined by reversing the order of the geometric product: ~(a*b) == ~b * ~a. For a k-vector a, the reversion is reversion_sign(k)*a where the sign is given by $(-1)^{k(k - 1)/2}$.

See also involution and clifford_conj.

GeometricAlgebra.scalar_prod — Method

a ⊙ b
scalar_prod(a, b) -> Number

Scalar part of the multivector product a*b.

GeometricAlgebra.wedge — Method

a ∧ b
wedge(a, b)

Wedge product of multivectors, a.k.a. the outer, exterior or alternating product.

This is a grade-raising operation, equivalent to graded_prod(+, a, b). If a and b are of grades $p$ and $q$ respectively, then a ∧ b is the grade $p + q$ part of a*b.

GeometricAlgebra.matrix_repr — Function

matrix_repr(a::AbstractMultivector, k=0:dim)

Matrix representation of the grade k parts of a multivector.

By default, the full $2^d × 2^d$ linear representation is used in $d$ dimensions. Smaller representations can be used for elements in

the even subalgebra, k=0:2:dim

the scalar-pseudoscalar subalgebra, k=(0, dim)

by restricting k to those grades.

Examples

julia> @basis 2
[ Info: Defined basis blades v, v1, v2, v12

julia> matrix_repr(v + 7v12)
4×4 Matrix{Int64}:
 1   0  0  -7
 0   1  7   0
 0  -7  1   0
 7   0  0   1

julia> matrix_repr(v + 7v12, (0, 2))
2×2 Matrix{Int64}:
 1  -7
 7   1

julia> matrix_repr(v1*v2) == matrix_repr(v1)matrix_repr(v2)
true

Metric Signatures

GeometricAlgebra.Cl — Type

Cl(p, q=0, r=0)

Metric signature where p, q and r is the number of basis vectors of norm +1, -1 and 0, respectively.

Examples

julia> basis(Cl(1,3))
4-element Vector{BasisBlade{Cl(1,3), 1, Int64}}:
 v1
 v2
 v3
 v4

julia> ans .^ 2
4-element Vector{BasisBlade{Cl(1,3), 0, Int64}}:
  1
 -1
 -1
 -1

GeometricAlgebra.basis — Function

basis(sig, grade=1)

Vector of basis blades of specified grade(s) for the geometric algebra defined by the metric signature sig.

See also @basis and @basisall.

Examples

julia> basis(3)
3-element Vector{BasisBlade{3, 1, Int64}}:
 v1
 v2
 v3

julia> basis("-+++", 0:2:4)
8-element Vector{BasisBlade{⟨-+++⟩, _A, Int64} where _A}:
 1
 v12
 v13
 v23
 v14
 v24
 v34
 v1234

julia> basis(Cl(1,3), 2)
6-element Vector{BasisBlade{Cl(1,3), 2, Int64}}:
 v12
 v13
 v23
 v14
 v24
 v34

GeometricAlgebra.cayleytable — Method

cayleytable(sig, op=*)
cayleytable(objs, op=*)

Display a multivector multiplication table.

The first argument may be a metric signature or any vector of objects which can be combined with the binary operator op.

The keyword argument title sets the contents of the top-left cell.

Examples

julia> cayleytable(3)
 (↓) * (→) │    1 │   v1     v2    v3 │  v12    v13   v23 │ v123
───────────┼──────┼───────────────────┼───────────────────┼──────
         1 │    1 │   v1     v2    v3 │  v12    v13   v23 │ v123
───────────┼──────┼───────────────────┼───────────────────┼──────
        v1 │   v1 │    1    v12   v13 │   v2     v3  v123 │  v23
        v2 │   v2 │ -v12      1   v23 │  -v1  -v123    v3 │ -v13
        v3 │   v3 │ -v13   -v23     1 │ v123    -v1   -v2 │  v12
───────────┼──────┼───────────────────┼───────────────────┼──────
       v12 │  v12 │  -v2     v1  v123 │   -1   -v23   v13 │  -v3
       v13 │  v13 │  -v3  -v123    v1 │  v23     -1  -v12 │   v2
       v23 │  v23 │ v123    -v3    v2 │ -v13    v12    -1 │  -v1
───────────┼──────┼───────────────────┼───────────────────┼──────
      v123 │ v123 │  v23   -v13   v12 │  -v3     v2   -v1 │   -1

julia> cayleytable(basis((t=-1, x=1, y=1, z=1), 2), ∧)
 (↓) ∧ (→) │   tx     ty    xy    tz     xz    yz
───────────┼──────────────────────────────────────
        tx │    0      0     0     0      0  txyz
        ty │    0      0     0     0  -txyz     0
        xy │    0      0     0  txyz      0     0
        tz │    0      0  txyz     0      0     0
        xz │    0  -txyz     0     0      0     0
        yz │ txyz      0     0     0      0     0

GeometricAlgebra.componentstype — Method

componentstype(sig, N, T)

Default array type used to store components of multivectors of signature sig. The resulting type should be able to store N components (in the case of a fixed-size array) of type T.

The fallback method returns Vector{T} for dimension(sig) <= 8, and SparseVector{T} otherwise.

GeometricAlgebra.ncomponents — Method

ncomponents(sig)
ncomponents(sig, k)

Dimension of (the grade-k subspace of) the geometric algebra of metric signature sig, viewed as a vector space.

If the dimension of the underlying vector space (see dimension) in $n$, then the algebra is $2^n$-dimensional, and its grade-$k$ subspace $\binom{n}{k}$-dimensional.

GeometricAlgebra.show_basis_blade — Method

show_basis_blade(io, sig, indices::Vector{Int})

Show the basis blade with unit vectors in indices for the geometric algebra defined by sig. Methods dispatching on sig should be added to customise basis blade labels for particular algebras.

Examples

julia> GeometricAlgebra.show_basis_blade(stdout, (1, 1, 1), [1, 3])
v13

julia> using GeometricAlgebra: subscript

julia> GeometricAlgebra.show_basis_blade(io, sig, indices) = print(io, join("𝒆".*subscript.(indices), "∧"))

julia> prod(basis(4))
BasisBlade{⟨++++⟩, 4, Int64} of grade 4:
 1 𝒆₁∧𝒆₂∧𝒆₃∧𝒆₄

GeometricAlgebra.show_signature — Method

show_signature(io, sig)

Pretty-print the metric signature sig.

This is used to display the metric signature type parameter in AbstractMultivector subtypes to reduce visual noise. Methods may optionally be added for user-defined metric signatures, in a similar fashion to Base.show.

Examples

julia> sig = (+1,-1,-1,-1)
(1, -1, -1, -1)

julia> GeometricAlgebra.show_signature(stdout, sig)
⟨+---⟩

julia> BasisBlade{sig}
BasisBlade{⟨+---⟩} (pretty-printed BasisBlade{(1, -1, -1, -1)})

GeometricAlgebra.@basis — Macro

@basis sig

Populate namespace with basis blades of every grade in the geometric algebra with metric signature sig.

Warning

This defines $2^n$ variables for an $n$ dimensional signature.

See also @basisall.

Examples

julia> @basis 3
[ Info: Defined basis blades v, v1, v2, v3, v12, v13, v23, v123

julia> 1v2 + 3v12
8-component Multivector{3, 0:3, Vector{Int64}}:
 1 v2
 3 v12

GeometricAlgebra.@basisall — Macro

@basisall sig

Similarly to @basis, populate namespace with basis blades, but include all permutations of each blade (e.g., v21 as well as v12).

Warning

This defines more than $2^n$ variables for an $n$ dimensional signature!

Examples

julia> @basisall (+1,-1)
[ Info: Defined basis blades v, v1, v2, v12, v21

julia> v12 == -v21
true

Internals

GeometricAlgebra.BitPermutations — Type

BitPermutations{T}(n)

Infinite iterator returning all unsigned integers of type T, in ascending order, for which Base.count_ones is n.

GeometricAlgebra.bits_of_grade — Method

bits_of_grade(k[, dim])

Generate basis blade bits of grade k in ascending order. Yields all basis blades in the dimension dim, if given, otherwise iterates indefinitely.

Examples

julia> GeometricAlgebra.bits_of_grade(2, 4) .|> UInt8 .|> bitstring
6-element Vector{String}:
 "00000011"
 "00000101"
 "00000110"
 "00001001"
 "00001010"
 "00001100"

GeometricAlgebra.bits_to_indices — Method

bits_to_indices(bits)

Return the positions of the ones in the unsigned integer bits.

Used to convert between representations of a unit blade. Inverse of indices_to_bits.

Examples

julia> GeometricAlgebra.bits_to_indices(0b1001101)
4-element Vector{Int64}:
 1
 3
 4
 7

GeometricAlgebra.componentbits — Method

componentbits(n, k)
componentbits(::Val{N}, ::Val{K})

Vector of bits corresponding to components of an n-dimensional Multivector of grade(s) k.

Bits are ordered first by grade (count_ones), then lexicographically (in ascending numerical order).

Passing Val arguments calls a faster, memoized method.

Examples

julia> componentbits(4, 2) .|> UInt8 .|> bitstring
6-element Vector{String}:
 "00000011"
 "00000101"
 "00000110"
 "00001001"
 "00001010"
 "00001100"

julia> componentbits(3, 0:3) .|> UInt8 .|> bitstring
8-element Vector{String}:
 "00000000"
 "00000001"
 "00000010"
 "00000100"
 "00000011"
 "00000101"
 "00000110"
 "00000111"

GeometricAlgebra.factor_from_squares — Method

factor_from_squares(sig, bits::Unsigned)

Compute the overall factor arising from the geometric product between repeated basis vectors.

GeometricAlgebra.geometric_prod_bits — Method

geometric_prod_bits(sig, a::Unsigned, b::Unsigned)

Compute the geometric product between unit blades. Returns a tuple of the overall scalar factor and the resulting unit blade.

GeometricAlgebra.indices_to_bits — Function

indices_to_bits(indices, T=UInt)

Create unsigned integer with bits at the positions given in the vector indices.

Used to convert between representations of a unit blade. Inverse of bits_to_indices.

Warning

Only correct if maximum(indices) does not exceed number of bits in T.

Examples

julia> GeometricAlgebra.indices_to_bits([1, 2, 5], UInt8) |> bitstring
"00010011"

GeometricAlgebra.next_bit_permutation — Method

Return the smallest uint larger than the one given which has the same number of binary ones. Algorithm is Gosper’s hack.

Examples

julia> GeometricAlgebra.next_bit_permutation(0b1011) |> bitstring
"00001101"

GeometricAlgebra.reversion_sign — Method

reversion_sign(k) = mod(k, 4) <= 1 ? +1 : -1

Sign from reversing a $k$-vector.

GeometricAlgebra.sign_from_swaps — Method

sign_from_swaps(a::Unsigned, b::Unsigned)

Compute sign flips of blade product due to transposing basis vectors into sorted order. (The full sign of the product will also depend on the basis norms.)

GeometricAlgebra.show_blade — Method

Display blade with parentheses surrounding coefficient if necessary.

Example

julia> GeometricAlgebra.show_blade(stdout, BasisBlade{(x=1,)}(0b1 => 1 + im))
(1+1im) x

GeometricAlgebra.show_multivector — Method

Display multivector components in a column or inline, optionally grouping by grade.

Parameters:

inline::Bool: print on one line (default true)

groupgrades::Bool: visually group components by grade (default true). If inline, draws parentheses around parts of each grade; if multiline, draw each grade on its own line

showzeros::Bool: whether to omit zero components from display

indent::Integer: indent amount

Examples

julia> a = Multivector{2,0:2}((1:4) .^ 2);

julia> GeometricAlgebra.show_multivector(stdout, a; inline=false, groupgrades=false)
 1 v
 4 v1
 9 v2
16 v12

julia> GeometricAlgebra.show_multivector(stdout, a; inline=false, groupgrades=true)
1
4 v1 + 9 v2
16 v12

julia> GeometricAlgebra.show_multivector(stdout, a; inline=true, groupgrades=true)
(1) + (4 v1 + 9 v2) + (16 v12)

julia> GeometricAlgebra.show_multivector(stdout, a; inline=true, groupgrades=false)
1 + 4 v1 + 9 v2 + 16 v12

GeometricAlgebra.symbolic_optim — Method

symbolic_optim(f, a, b, ...)

Trace evaluation of f(a, b, ...)::CompositeMultivector on symbolic versions of each AbstractMultivector instance or type a, b, ..., returning an expression suitable as the body of a @generated function.

Important

The names of the generated function arguments must be "a", "b", "c", etc, as these are the names used in the expression retuned by symbolic_optim.

If use_symbolic_optim(sig) returns false, the function body simply calls f(a, b, ...).

Examples

using MacroTools: prettify
u, v = Multivector.(basis(2))
ex = GeometricAlgebra.symbolic_optim(*, u, v) |> prettify

# output
:(let a = components(a), b = components(b)
      comps = create_array(Vector{Int64}, Int64, Val{1}(), Val{(2,)}(), a[1] * b[1] + a[2] * b[2], a[1] * b[2] + (-1 * a[2]) * b[1])
      (Multivector{2, 0:2:2})(comps)
  end)

GeometricAlgebra.use_symbolic_optim — Method

use_symbolic_optim(sig) -> Bool

Whether to use symbolic code generation to optimize operations in algebras of metric signature sig.

By default, this is enabled if dimension(sig) ≤ 8 as a heuristic (in many dimensions, algebraic expressions may become too unwieldy).

D02Leo Dorst, "The Inner Products of Geometric Algebra", 2002. doi:10.1007/978-1-4612-0089-5_2

D02Leo Dorst, "The Inner Products of Geometric Algebra", 2002. doi:10.1007/978-1-4612-0089-5_2