Reference

GeometricAlgebra.GeometricAlgebra — Module

GeometricAlgebra

Implementation of multivectors, or $k$-vectors, as in geometric (Clifford) algebra.

GeometricAlgebra.AbstractMultivector — Type

AbstractMultivector{Sig}

Supertype of all elements in the geometric algebra defined by the metric signature Sig.

Subtypes

                   AbstractMultivector{Sig}
                     /                  \
  HomogeneousMultivector{Sig,K}      MixedMultivector{Sig,S}
       /                \                             
Blade{Sig,K,T}   Multivector{Sig,K,S}                
                                                   
                 ╰────── CompositeMultivector{Sig,S} ──────╯

Blade: a scalar multiple of a wedge product of orthogonal basis vectors.

Multivector: a homogeneous multivector; a sum of same-grade blades.

MixedMultivector: an inhomogeneous multivector. All elements in a geometric algebra can be represented as this type (though not most efficiently).

Note

The mathematical definition of a $k$-blade is the wedge product of $k$ vectors, not necessarily basis vectors. Thus, not all $k$-blades are representable as a Blade, but are always representable as a sum of Blades, or as a Multivector.

Type Parameters

Sig: The metric signature which defines the geometric algebra. This can be any all-bits value which satisfies the metric signature interface. For example, (1, 1, 1) or EuclideanSignature(3) both define the standard geometric algebra of $ℝ^3$.

T: The numerical type of the coefficient of a Blade.

K: An Int specifying the grade of a HomogeneousMultivector.

S: The storage type of the components of a CompositeMultivector. This is assumed to be mutable, and is usually a subtype of Vector, MVector or SparseVector.

GeometricAlgebra.BitPermutations — Type

BitPermutations{T}(n)

Infinite iterator returning all unsigned integers of type T, in ascending order, for which Base.count_ones is n.

GeometricAlgebra.Blade — Type

Blade{Sig,K,T} <: HomogeneousMultivector{Sig,K}

A blade of grade K ∈ ℕ with basis blade bits and scalar coefficient of type T.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K: Grade of the blade, equal to count_ones(bits), retrieved with grade().

T: Numerical type of the scalar coefficient, retrieved with eltype().

GeometricAlgebra.HomogeneousMultivector — Type

HomogeneousMultivector{Sig,K} <: AbstractMultivector{Sig}

Abstract supertype of Blade and Multivector.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K: Grade of the blade or multivector, retrieved with grade().

GeometricAlgebra.MixedMultivector — Type

MixedMultivector{Sig,S} <: AbstractMultivector{Sig}

A (possibly inhomogeneous) multivector.

All elements of a geometric algebra are representable as a MixedMultivector.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

S: Storage type of the multivector components, generally a subtype of AbstractVector.

GeometricAlgebra.Multivector — Type

Multivector{Sig,K,S} <: HomogeneousMultivector{Sig,K}

A homogeneous multivector of grade K ∈ ℕ with storage type S.

Parameters

Sig: Metric signature defining the geometric algebra, retrieved with signature().

K: Grade of the multivector, retrieved with grade().

S: Storage type of the multivector components, generally a subtype of AbstractVector.

Base.eltype — Method

eltype(::AbstractMultivector)

The numerical type of the components of a multivector object or type.

GeometricAlgebra.bits_of_grade — Method

bits_of_grade(k[, dim])

Generate basis blade bits of grade k in ascending order. Yields all basis blades in the dimension dim, if given, otherwise iterate indefinitely.

Example

julia> GeometricAlgebra.bits_of_grade(2, 4) .|> UInt8 .|> bitstring
6-element Vector{String}:
 "00000011"
 "00000101"
 "00000110"
 "00001001"
 "00001010"
 "00001100"

GeometricAlgebra.bits_to_indices — Method

bits_to_indices(bits)

Return the positions of the ones in the unsigned integer bits.

Used to convert between representations of a unit blade. Inverse of indices_to_bits.

Examples

julia> GeometricAlgebra.bits_to_indices(0b1001101)
4-element Vector{Int64}:
 1
 3
 4
 7

GeometricAlgebra.bits_to_mmv_index — Method

bits_to_mmv_index(bits::Unsigned)

Convert a unit blade bits to a linear index for accessing components of a MixedMultivector.

GeometricAlgebra.bits_to_mv_index — Method

bits_to_mv_index(bits::Unsigned)

Convert a unit blade bits to a linear index for accessing components of a Multivector.

GeometricAlgebra.componentstype — Method

componentstype(sig, N, T)

Array type to use the components for multivectors of signature sig. The resulting type should be able to store N components (in the case of a fixed-size array) of element type T.

This is used when converting a Blade into a CompositeMultivector to determine the type of the components array.

GeometricAlgebra.dimension — Method

dimension(::AbstractMultivector)

The dimension of the underlying vector space of the geometric algebra. See ncomponents the number of independent components of a multivector.

GeometricAlgebra.factor_from_squares — Method

factor_from_squares(sig, bits::Unsigned)

Compute the overall factor arising from the geometric product between repeated basis vectors.

GeometricAlgebra.geometric_prod_bits — Method

geometric_prod_bits(sig, a::Unsigned, b::Unsigned)

Compute the geometric product between unit blades. Returns a tuple of the overall scalar factor and the resulting unit blade.

GeometricAlgebra.grade — Method

grade(::HomogeneousMultivector{Sig,K}) -> K

The grade of a homogeneous multivector (Blade or Multivector) object or type.

GeometricAlgebra.indices_to_bits — Method

indices_to_bits(indices)

Create unsigned integer with bits at the positions given in the vector indices.

Used to convert between representations of a unit blade. Inverse of bits_to_indices.

Warning

Produces incorrect results if elements of indices are greater than the number of bits in bits_scalar() <: Unsigned.

Examples

julia> GeometricAlgebra.indices_to_bits([1, 2, 5]) |> UInt16 |> bitstring
"0000000000010011"

GeometricAlgebra.mmv_index_to_bits — Method

mmv_index_to_bits(ith, dim)

Convert a linear index for MixedMultivector components into the corresponding unit blade. Uses a lookup table GeometricAlgebra.MULTIVECTOR_INDICES for speed.

GeometricAlgebra.mv_index_to_bits — Method

mv_index_to_bits(ith, k)

Convert a linear index for grade-k Multivector components into the corresponding unit blade.

GeometricAlgebra.ncomponents — Method

ncomponents(::CompositeMultivector)

Number of independent components of a multivector object or type.

In $n$ dimensions, this is $\binom{n}{k}$ for a Multivector and $2^n$ for a MixedMultivector.

GeometricAlgebra.next_bit_permutation — Method

Return the smallest uint larger than the one given which has the same number of binary ones. Algorithm is Gosper's hack.

julia> GeometricAlgebra.next_bit_permutation(0b1011) |> bitstring
"00001101"

GeometricAlgebra.show_blade — Method

Display blade with parentheses surrounding coefficient if necessary.

julia> GeometricAlgebra.show_blade(stdout, Blade{(x=1,)}(0b1 => 1 + im))
(1+1im) v1

GeometricAlgebra.show_mixedmultivector — Method

Display an inhomogeneous MixedMultivector with each grade on a new line.

GeometricAlgebra.show_multivector — Method

Display homogeneous multivector components as a column of blades, with coefficients and blades aligned using the native alignment mechanism.

julia> a = Multivector{(1,1,1),1}([1e3, 1, 1e-3]);

julia> GeometricAlgebra.show_multivector(stdout, a)
1000.0   v1
   1.0   v2
   0.001 v3

GeometricAlgebra.sign_from_swaps — Method

Compute sign flips of blade product due to transposing basis vectors. (The full sign of the product will also depend on the basis norms.)

GeometricAlgebra.signature — Method

signature(::AbstractMultivector{Sig}) -> Sig

The metric signature type parameter of the multivector object or type.